requirements.txt
streamlit
pytesseract 
pillow
python-dotenv
langchain
PyPDF2
chromadb
google-generativeai
scrapegraph-py
crawl4ai
langgraph
langchain-google-genai
fastapi
uvicorn
fastapi
uvicorn
pydantic
crewai
crewai_tools
python-dotenv
python-multipart
python-docx
pymupdf
agentops
markdown-it-py
weasyprint
mdit_py_plugins
websockets
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]